Refactor superbase to superbase_user
