+++ components/ProfileSettings/EditProfile.tsx
@@ -1,7 +1,6 @@
 import { View, Text, StyleSheet, SafeAreaView, Alert, ActivityIndicator, TouchableOpacity } from 'react-native'
 import React, { useEffect, useState } from 'react'
-import { getCurrentUser, handleUserDeletion } from '@/lib/supabase'
-import { User } from '@/types/types'
+import { getCurrentUser, handleUserDeletion } from '@/lib/supabase_user'
 import CustomButton from '../CustomButtom'
 import { useGlobalContext } from '@/context/Context'
 import { router } from 'expo-router'
